// src/components/TradingView/charting_library.d.ts
export interface Bar {
  /** Bar time.
   * Amount of **milliseconds** since Unix epoch start in **UTC** timezone.
   * `time` for daily, weekly, and monthly bars is expected to be a trading day (not session start day) at 00:00 UTC.
   * The library adjusts time according to `session` from {@link LibrarySymbolInfo}.
   */
  time: number
  /** Opening price */
  open: number
  /** High price */
  high: number
  /** Low price */
  low: number
  /** Closing price */
  close: number
  /** Trading Volume */
  volume?: number
}